Fixed join command in dm, and removed old logging

+++ src/Commands.ts
@@ -4,7 +4,6 @@ import config from "./config";
 import NotABot from "./Guards/NotABot";
 import Main from "./Main";
 import { userJoined } from "./service";
-import logger from "./utils/logger";
 
 @Discord(config.prefix)
 abstract class Commands {
@@ -22,11 +21,10 @@ abstract class Commands {
   @Guard(NotABot)
   join(command: CommandMessage): void {
     const { joinCode } = command.args;
-    logger.debug(
-      `User joined (${joinCode}, "discord", ${command.author.id}, ${command.guild.id})`
-    );
-    userJoined(joinCode, command.author.id, command.guild.id, true);
-    command.delete().then();
+    userJoined(joinCode, command.author.id, command.guild?.id, true);
+    if (command.channel.type !== "dm") {
+      command.delete().then();
+    }
   }
 }
